Added delete functionality to surveys

+++ Enodo/Capstone_Project/Controllers/API/SurveyController.cs
@@ -4,9 +4,11 @@ using System.Linq;
 using System.Data.Entity;
 using System.Web;
 using System.Web.Http;
+using System.Web.Http.Results;
 using AutoMapper;
 using Capstone_Project.Dtos;
 using Capstone_Project.Models;
+using Microsoft.AspNet.Identity;
 
 namespace Capstone_Project.Controllers.api
 {
@@ -78,14 +80,34 @@ namespace Capstone_Project.Controllers.api
         public IHttpActionResult DeleteSurvey(int id)
         {
             var surveyInDb = _context.Surveys.SingleOrDefault(c => c.Id == id);
+            var optionsInDb = _context.Options.Where(o => o.SurveyId == id);
+            var surveyResultsInDb = _context.SurveyResultsSet.Where(o => o.SurveyId == id);
+
+            var currentUserId = HttpContext.Current.User.Identity.GetUserId();
+            var currentUser = _context.Users.Single(u => u.Id == currentUserId);
 
             if (surveyInDb == null)
                 return NotFound();
 
-            _context.Surveys.Remove(surveyInDb);
-            _context.SaveChanges();
+            if (currentUser.UserName == surveyInDb.Owner)
+            {
+                foreach (var option in optionsInDb)
+                {
+                    _context.Options.Remove(option);
+                }
 
-            return Ok();
+                foreach (var surveyResult in surveyResultsInDb)
+                {
+                    _context.SurveyResultsSet.Remove(surveyResult);
+                }
+
+                _context.Surveys.Remove(surveyInDb);
+                _context.SaveChanges();
+
+                return Ok();
+            }
+
+            return null;
         }
     }
 }
